#0: Add env var for skipping core resets on init
For cases where running right after a tt-smi reset, there's a chance
that L1 comes up in a random state that tricks metal into thinking
there's kernels left running. In that case it may try to force exit
them, but if a kernel is not actually running this can result in a hang,
so add an env var (TT_METAL_SKIP_RESET_CORES_ON_INIT) for this specific
case.TT_METAL_SKIP_RESET_CORES_ON_INIT) for this specific case.